requirements.txt
asgiref==3.8.1
colorama==0.4.6
coverage==7.6.1
Django==5.0.6
django-appconf==1.0.6
django-db-logger==0.1.13
django-decouple==2.1
django-tinymce==4.1.0
django_compressor==4.5
fontawesomefree==6.5.1
gunicorn==22.0.0
iniconfig==2.0.0
packaging==24.1
pillow==10.3.0
pluggy==1.5.0
psycopg2==2.9.9
pytz==2024.2
rcssmin==1.1.2
rjsmin==1.2.2
setuptools==75.1.0
six==1.16.0
sqlparse==0.5.0
tzdata==2024.1
wheel==0.44.0
whitenoise==6.7.0
